requirements.txt
beautifulsoup4==4.3.2
defusedxml==0.4.1
Django==1.7.2
django-allauth==0.18.0
django-braces==1.4.0
django-crispy-forms==1.4.0
django-dynamic-fixture==1.7.0
django-localflavor==1.0
django-pagination==1.0.7
django-secure==1.0
django-webtest==1.7.7
factory-boy==2.4.1
oauthlib==0.7.2
Pillow==2.7.0
psycopg2==2.5.4
pygeocoder==1.2.5
python3-openid==3.0.5
requests==2.5.1
requests-oauthlib==0.4.2
selenium==2.42.1
six==1.7.3
stripe==1.18.0
waitress==0.8.9
WebOb==1.4
WebTest==2.0.15
